play only passes integer length now.

+++ out.py
@@ -24,4 +24,4 @@ def wav_write(fname,stream,rate=44100,width=2,channels=1):
 
 def play(stream,duration,frac=0.0,rate=44100):
     with AudioIO(True) as player:
-        player.play(buffer_stream(stream,rate*duration,frac),rate=rate) 
+        player.play(buffer_stream(stream,int(rate*duration),frac),rate=rate)
